+++ HISTORY.md
@@ -9,6 +9,7 @@
 
 * Minor S3 optimizations and error message fixes (PR #462)
 * Changed S3 default config so that AWS S3 just works (PR #455)
+* Got rid of special S3 "directory objects"
 
 ## Bug Fixes
 
@@ -35,12 +36,16 @@
 ### C API
 
 * `tiledb_query_finalize` must **always** be called before `tiledb_query_free`.
+* Removed `tiledb_vfs_move` and added `tiledb_vfs_move_file` and `tiledb_vfs_move_dir` instead.
+* Removed `force` argument from `tiledb_vfs_move_*` and `tiledb_object_move`.
 
 ### C++ API
 * Fixes with `Array::max_buffer_elements` and `Query::result_buffer_elements` to comply with the API docs. `pair.first` is the number of elements of the offsets buffer. If `pair.first` is 0, it is a fixed-sized attribute or coordinates.
 * `std::array<T, N>` is backed by a `char` tiledb attribute since the size is not guaranteed.
 * `Attribute::set_cell_val_num()` is deprecated. This is now deduced from the Attribute type.
 * Headers have the `tiledb_cpp_api_` prefix removed. For example, the include is now `#include <tiledb/attribute.h>`
+* Removed `VFS::move` and added `VFS::move_file` and `VFS::move_dir` instead.
+* Removed `force` argument from `VFS::move_*` and `Object::move`.
 
 # TileDB v1.2.1 Release Notes